+++ examples/resources/nxos_vpc_domain/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_vpc_domain.example "sys/vpc/inst/dom"