+++ docs/guides/changelog.md
@@ -15,6 +15,7 @@ description: |-
 - Add `access_list_name` attribute to `nxos_ipv4_access_list_policy_egress_interface` resource and data source
 - BREAKING CHANGE: Remove  `nxos_ipv4_access_list_policy_egress_interface_instance` resource and data source
 - Add `nxos_vpc_instance` resource and data source
+- Add `nxos_vpc_domain` resource and data source
 
 ## 0.3.23